+++ test/libsyncengine/syncpal/testoperationprocessor.cpp
@@ -29,7 +29,6 @@
 #include "mocks/libcommonserver/db/mockdb.h"
 
 #include "test_utility/testhelpers.h"
-#include "test_classes/testsituationgenerator.h"
 #include "update_detection/file_system_observer/filesystemobserverworker.h"
 
 #include <cstdlib>
@@ -113,7 +112,7 @@ void TestOperationProcessor::testIsPseudoConflict() {
     // .
     // └── A
     //     └── AA
-    TestSituationGenerator situationGenerator(_syncPal);
+    SituationGenerator situationGenerator(_syncPal);
     situationGenerator.generateInitialSituation(R"({"a":{"aa":1}})");
 
     const auto lNodeA = situationGenerator.getNode(ReplicaSide::Local, "a");
