--- test/test_classes/testsituationgeneratortest.h
@@ -1,45 +1,0 @@
-/*
- * Infomaniak kDrive - Desktop
- * Copyright (C) 2023-2026 Infomaniak Network SA
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#pragma once
-
-#include "testincludes.h"
-#include "test_classes/testsituationgenerator.h"
-
-namespace KDC {
-
-class TestSituationGeneratorTest : public CppUnit::TestFixture, public TestBase {
-    public:
-        CPPUNIT_TEST_SUITE(TestSituationGeneratorTest);
-        //  CPPUNIT_TEST(testLegacyFormat);
-        CPPUNIT_TEST(testExtendedFormat);
-        CPPUNIT_TEST_SUITE_END();
-
-    public:
-        void setUp() override;
-        void tearDown() override;
-
-    protected:
-        void testLegacyFormat();
-        void testExtendedFormat();
-
-    private:
-        TestSituationGenerator _gen;
-};
-
-} // namespace KDC
